+++ board/digi/common/hwid.h
@@ -54,7 +54,7 @@ void board_hwid_update(bool is_fuse);
 int board_hwid_fuse_lock(void);
 void fdt_fixup_hwid(void *fdt, const struct digi_hwid *hwid);
 void fdt_fixup_fuse_hwid(void *fdt);
-u32 hwid_get_ramsize(const struct digi_hwid *hwid);
+u64 hwid_get_ramsize(const struct digi_hwid *hwid);
 void hwid_get_macs(const struct digi_hwid *hwid);
 void hwid_get_mac_pool(const struct digi_hwid *hwid, uint8_t *mac);
 void hwid_get_serial_number(const struct digi_hwid *hwid);
